refactor bottom up fuser
